gdb/comrv: Fix bug #28, duplicate breakpoints set in overlay code
Bug #28 describes a situation where we have a situation like this:

  volatile int global_var = 0;

  void
  some_overlay_function ()
  {
    global_var = 0;
  }			/* Breakpoint 1.  */

  void
  some_function ()
  {
    some_overlay_function ();
    global_var = 1;	/* Breakpoint 2.  */
  }

The user creates 'Breakpoint 1' and 'Breakpoint 2', then continues and
hits breakpoint 1.  At this point the user uses 'finish' and expects
to end up at the location of 'Breakpoint 2'.

Now GDB should only insert a single breakpoint at the address of
'Breakpoint 2', but we observe that GDB was actually inserting two
breakpoints at this address.

First, the remote protocol specification is clear that this is
allowed, and any target that complains about this is a broken
target.... but... sometimes we have to work with broken targets, and
as GDB tries to avoid inserting duplicate breakpoints, lets fix GDB.

The problem is that the breakpoint places by the user is assigned a
different section to the breakpoint created by 'finish'.  As a result
of this the existing overlay code in GDB decides these two breakpoints
are different and inserts both.

As the ComRV overlay code doesn't make use of the sections to manage
overlays this check is invalid for us.  The different section is
meaningless and should not cause an extra breakpoint to be inserted.
So this commit removes this check.