gdb/comrv: Fix bug #30, memory corruption after comrv reset
A case was discovered where after calling the comrv reset method
memory corruption was experienced.  What happens is:

  1. Insert software breakpoint into an overlay.  The current memory
  contents are backed up within the breakpoint, and a software
  breakpoint is inserted.

  2. At some later point ComRV replaces the current overlay with a new
  one.

  3. We try to remove the breakpoint inserted in step #1.  To do this
  we:

    (a) Read the new memory contents,
    (b) Remove the breakpoint, this writes back the old memory
    contents (as read in step #1), which is incorrect.
    (c) Write back the contents read in #a.

The problem with this is that in #a, when we read memory GDB spots
that there is a breakpoint at this location and instead of returning
the actual memory contents the contents as cached in #1 are returned.

Then when we write back what we thought were the current contents in
step #c, we are actually writing back the old contents from #1.

To fix this we need to disable the breakpoint shadow cache during the
step #a.  We do this by temporarily setting the breakpoints shadow
cache size to 0.

To further optimise things, when reading the current contents in #a,
the contents are placed directly into the breakpoints shadow contents
buffer.  Now when we remove the breakpoint we write back the actual
current memory contents.  With this done step #c is no longer needed.