gdb/comrv: additional bug #30 fixes, disable z0/Z0 packets
Using z0/Z0 packets with overlays will not work in one specific case:

  1. Breakpoint is inserted with z0 into overlay group 1.

  2. Overlay group 1 is replaced with overlay group 2.

  3. We try to remove the breakpoint that was inserted into overlay
  group 1.

When GDB sends the Z0 packet the remote will write the original
contents back in at the breakpoint address.  Unfortunately this
content is from overlay group 1, and so overlay group 2 becomes
corrupted.

The previous attempt to fix bug #30 works fine if we are using GDB's
memory breakpoints.  In this case we effectively disable writing back
of the old memory contents.