gdb/riscv: select rv32 target description when appropriate
Currently, when RISC-V GDB starts without an executable an RV64 target
description will always be selected.

This is even true when the user has tried to explicitly set the
architecture to RV32.   This is a current GDB session:

  $ gdb
  (gdb) target remote ......
  (gdb) set arch riscv:rv32
  (gdb) print sizeof ($pc)
  $1 = 8

After this commit we instead see this:

  $ gdb
  (gdb) target remote ......
  (gdb) set arch riscv:rv32
  (gdb) print sizeof ($pc)
  $1 = 4

GDB has now selected the RV32 description as the default in order to
reflect the user specified architecture.