gdb: include target section table in 'maint info sections' output
The output of 'maint info sections' shows all the sections from two
sources:

  1. The current executable, and
  2. The current core file.

Ideally I would like this to also include:

  1. Any additional symbol files that have been loaded,
  2. Some indication of which sections are in GDB's internal section
  tables.

This commit covers point #2 on the wish list.

I do wonder though if this information would be better moved into a
different command.  The current command is more about examining the
objects the user asked to load, while point #2 is really about a
different GDB internal data structure.

As this is a maintenance command I'm not going to worry about this too
much for now, I'm leaving this here as it helped me debug an issue,
but this might need more thought.