gdb/comrv: include all sections in section table when loading a symbol file
In this commit:

  commit 9ff47a38ec18a38126c8c09454ff43d31071ff8e
  Date:   Wed Sep 23 15:59:35 2020 +0000

      gdb/comrv: allow reading overlay storage directly from the ELF

GDB was changed so that all sections were tracked, not just the
allocatable sections.  Doing this allowed us to read the comrv storage
area from the ELF, which means we can now disassemble non-mapped
overlay functions.

Unfortunately, the change in the previous commit was only for the main
executable, that is the file passed on the GDB command line, or loaded
with the 'file' command.

This is unfortunate because the client is driving GDB through Eclipse,
and for some reason Eclipse is not using 'file' (or the MI
equivalent), but is instead using add-symbol-file.  When this means is
that GDB never actually has a main executable, instead there's just a
remote target with a symbol file loaded into memory.

And the above commit didn't cover additional symbol files.  So, when a
symbol file is added only the allocatable sections are added into the
section table.  This means that the comrv storage sections are not
available to be read from when needed, and we can't disassemble
non-mapped overlay functions.

With this commit we now add all sections to the section table.  To do
this we must first add all sections into the objfile data structure.

So that we don't break existing code I have added a flag to the
obj_section structure, we can then mark those sections that are
"extra" (i.e. would not normally have been added).  This allows all
of the existing users of this section list to filter out the sections
that they previously would not have seen.

Only in the once place where sections are copied into GDB's internal
section table do we now loop over all sections, including the extra
sections, thus ensuring the section table contains all required
sections.

Finally, I had to implement get_section_table for the remote target.
Normally when a user specifies an executable and connects to a remote
target we would have a remote target above an exec target, the exec
target would take care of handling the get_section_table request.

However, in our strange Eclipse use case there is no exec target,
there's just a remote target.  However, the per program space section
table is still valid and available, so just add a method to the remote
target that provides this information.