gnulib: work around issue with gnulib on RHEL6
On RHEL6 (and derived variants) we observe a build issue with gnulib,
where we end up not understanding the macro __THROWNL.

This macro is expected to be defined in the system cdefs.h file,
however, if it is not the gnulib provides its own definition.  The
problem is that the local definition is only provided if the system
cdefs.h file is considered to be "too old".

The way that we check if the system cdefs.h file is "too old" is by
checking to see if __glibc_likely is defined or not.  In upstream
glibc the __glibc_likely macro was added after __THROWNL was added.

The problem is that on RHEL6 systems, the __glibc_likely macro appears
to have been back-ported to earlier versions of glibc, while the
__THROWNL macro has not.

As a result, if we try to compile on a suitably old RH6 system, we do
have __glibc_likely, but we don't have __THROWNL, and so the
compilation fails.

We fix this by checking that both __THROWNL and __glibc_likely exist.
If either of them don't exist then we pull in the local cdefs.h file.

gnulib/ChangeLog:

	* import/libc-config.h: Check for __THROWNL not existing, just as
	we do for __glibc_likely, this fixes a Redhat 6 issue.